fix: review suggestion
