+++ docs/src/components/drag-overlay.md
@@ -1,43 +1,50 @@
+---
+outline: deep
+---
+
 # DragOverlay 拖拽浮层
 
 一个提供拖拽上传能力的组件，通过自定义指令 `v-drag-aware` 和一个纯展示的浮层组件 `<tr-drag-overlay>` 协同工作。
-
-## 何时使用
-
-当您需要为一个特定区域或整个页面添加文件拖拽上传功能，并在拖拽时显示一个全局浮层时，可以使用本组件。
-
-## 组件构成
 
 本功能由两部分组成：
 
 -   `v-drag-aware`: 一个自定义 Vue 指令，负责监听和处理DOM元素的拖拽事件。
 -   `<tr-drag-overlay>`: 一个纯展示组件，根据传入的 `is-dragging` prop 显示或隐藏一个全屏的拖拽浮层。
 
-## 基本用法
+## 代码示例
+
+### 基本用法
 
 将 `v-drag-aware` 指令附加到任何你希望响应拖拽的元素上。同时，在页面中放置一个 `<tr-drag-overlay>` 组件，并通过一个状态变量将其 `is-dragging` prop 与指令的状态同步。
 
-
 <demo vue="../../demos/drag-overlay/basic.vue" />
 
-## 综合案例
+### 自定义拖拽层
 
-<demo vue="../../demos/drag-overlay/container-integration.vue" />
+<demo vue="../../demos/drag-overlay/custom-overlay.vue" />
 
-## Attributes
+### 状态禁用
 
-### v-drag-aware 指令绑定对象
+<demo vue="../../demos/drag-overlay/disabled.vue" />
+
+## API
+
+### Attributes
+
+**v-drag-aware** 指令传递的参数
 
 | 名称             | 类型                                   | 说明                                             |
 | ---------------- | -------------------------------------- | ------------------------------------------------ |
 | onStateChange    | `(isDragging: boolean) => void`        | **必须**。当拖拽状态改变时触发的回调。         |
 | onFilesDropped   | `(files: File[]) => void`              | **必须**。当符合要求的文件被放下时触发的回调。 |
-| onFilesRejected  | `(rejection: FileRejection) => void`   | 当文件因类型不符被拒绝时触发的回调。             |
+| onFilesRejected  | `(rejection: FileRejection) => void`   | 当文件因不符条件被拒绝时触发的回调。             |
 | accept           | `string`                               | 可选。接受的文件类型，同 `<input>` 的 `accept` 属性。 |
 | multiple         | `boolean`                              | 可选，默认为 `true`。是否允许上传多个文件。       |
 | disabled         | `boolean`                              | 可选，默认为 `false`。是否禁用拖拽功能。          |
+| maxSize          | `number`                                | 可选。允许上传的最大文件大小，单位为字节。默认为 10 MB |
+| maxFiles         | `number`                               | 可选。允许上传的最大文件数量。默认为 3。         |
 
-### TrDragUploadWrapper Props
+### Props
 
 | 名称               | 类型      | 默认值 | 说明                                     |
 | ------------------ | --------- | ------ | ---------------------------------------- |
@@ -47,11 +54,26 @@
 | targetRect | `DOMRect \| null`| `null`   | 目标元素的 DOMRect，用于定位覆盖层。 |
 | fullscreen | `boolean`| `false`   | 是否全屏模式，控制覆盖层的边框显示。 |
 
-
-## Slots
-
-### TrDragUploadWrapper Slots
+### Slots
 
 | 名称      | 说明           |
 | --------- | -------------- |
 | overlay   | 自定义浮层内容。 |
+
+### Types
+
+**FileRejection**
+
+```typeScript
+export interface FileRejection {
+  files: File[]
+  /**
+   * 拒绝原因
+   * @description 可能的值：
+   * - 'invalid-file-type'：文件类型不支持
+   * - 'invalid-file-size'：文件大小超出限制
+   * - 'invalid-file-count'：文件数量超出限制
+   */
+  reason: 'invalid-file-type' | 'invalid-file-size' | 'invalid-file-count'
+}
+```
