--- packages/components/src/demo/index.ts
@@ -1,3 +1,0 @@
-// Demo 目录的入口文件
-// 这个文件用于解决构建时的入口问题
-export {}
